Added preliminary "go to conversation" support.
Really, what I want is a GNU Social-style "In reply to <USER>" link.
Unfortunately, it looks like I can't easily get the user's display name
from Atom data, and the page linked to by in-reply-to data isn't
accessible by JS.  So until I have a better answer, we'll just link to
the conversation (as given by the conversation tag) locally.