Corrected some overflow issues.
The userinfo block will now clip any text that goes too long (i.e. a
URL that's too long or a display name that is very long and has no
spaces).  The content area will do the same thing in case of a post
which contains long strings of unbreakable text, but that's a bit
sloppier of CSS to make it work.

What I'm saying is, I'm no CSS expert, and this entire stylesheet could
probably benefit from someone better at it giving it a once-over.