Added time-delayed text for a longer-than-expected load time.
Before, the "(in theory)" text was baked into the default loading text,
because that's what I had when I was testing things out and it felt
right.  Now that it seems to be working, I've changed it so that it
only shows up if the load takes longer than five seconds.  This can be
reconfigured at your leisure.