Added rel="noopener" to all outbound links.
And by "all outbound links", I more or less mean all links, as everything is an
outbound link from this widget.  Setting rel="noopener" should force browsers to
open everything in new processes, which should be what you'd want for safety and
security purposes if you're going to a different site and all.