Made sure only the feed's author data is used.
Turns out, if you repeat someone's posts, it winds up embedding their
author data in said repeated posts (so we could properly put avatars
and other data in place if we were displaying them).  That data comes
through as more author elements, so we can't safely just call
find("author") on the entire XML output (we would get all the authors
the user repeated in the last 20 posts or so).  Now we make sure we
ONLY get the author element that's an immediate child of the main feed
element.