Allowed HTML summaries for author data.
Apparently Mastodon stores author summaries as HTML data.  Fortunately,
they're nice enough to inform us of this fact with a type attribute.
To that end, if the summary comes back as HTML, gswidget will now
output it as such to the page.  If not, it'll force it to text.

I reserve the right to change this behavior to stripping out HTML
entirely if I deem it to be too dangerous, but the assumption here for
the time being is that the user is pointing this widget to known data
on a known server.

Also, Mastodon is no longer giving CORS errors when accessing the Atom
feed from scripts.  Yay!