Made the post list use the API.
As it stands, now it looks like it used to.  Sort of.  There's some new
layout issues now that the author summary is HTML.  I'll fix that later.

The next issue is to sanitize all this HTML we're getting, and then add
in the nifty new data I can get from the API, like recognizing boosts,
indicating to what a post is a reply, adding images and such, etc.