Cleaned up a bunch of CSS.
The biggest thing is that now the userinfo block isn't a horrible hack
of CSS floats.  Yay!  But this also does away with I think all of the
explicit references to the pixel dimensions of the iframe used in the
sample HTML file (which, in turn, is just what I use in the DoM's
implementation of it).  It also includes horizontal scrollbars on a
per-entry basis if the post does something weird that forces everything
to spill to the side.