Removed the user URL from the userinfo box.
Note that both the avatar and the username still link to it, of course.
Those make sense.  However, since the Mastodon API doesn't give me an
easy way to get a sort of "@username@instance" string to link off of,
just dumping the URL in its place doesn't really work the same way.