Changed entries to include user avatars and original boosted users.
Now each entry has a header including the user avatar and display name
of whoever posted it (important for boosts).  This change also
coincidentally makes the date format shorter (with the long version
available as a title popup).

With user avatars on each entry, the top part might be redone soon to
remove it and reorganize.