Cleaned up a bit of link style stuff.
Mentions are now boldfaced.  It looks like I can't get sufficient data
for replies in the API to do what I want, so this may be all I can do.

Also, the username link in the top block *itself* is what gets the
larger and bold styles, not its entire div.  That'll also make sense
later.