Changed fetch callbacks into Promises, also did a bunch of cleanup.
I decided the weird chaining of fetchAccountData and fetchStatuses was
sloppy and unintuitive, so now I'm using jQuery's get methods as proper
Promises and taking advantage of Promise.all to simplify things a lot.

In the process, though, I also found a lot more that ought to be cleaned
up, especially things like storing userData and postData as globals,
the fact that genericFetchError doesn't need to take a parameter, some
of the functions could use better names, and how showError was, in fact,
broken.