Separated the user-filled instance URL and user ID into variables.
Instead of having the user put in the URL to the API endpoint for the
account, now there's two variables, one for the instance URL and one for
the user's ID.  The actual accountUrl and statusesUrl used in-code are
constructed from these.

This is largely to allow for the upcoming code that'll be able to
properly find the parent toot and author to a reply toot.  Turns out the
best way to find those is to read them from the public timeline of the
current instance, but I'll get more into that in the commit message when
I make the extra fetches.

Incidentally, this change DOES mean anyone using this code will need to
update their variables, obviously.  Shouldn't be a tricky fix.