Moved user-configurable stuff into a config object.
In an attempt to clean stuff up a bit, the config variables are now in
their own convenient object, rather than just variables scattered
hither and yon with naught but comments to guide the user.  I mean,
this new object is just dumped in there as well, but at least it's more
on the organized side.  Kinda.

At time of commit, this hasn't been tested.