Changed to an only-update-what-changed model for displaying toots.
This one's a bit heftier than usual, as it changes the way Mastowidget
updates displayed toots from just wiping the entire list and starting
fresh on each refresh to only changing things if something updated on
the server (i.e. a new post was made, a post was deleted, a post was
edited).

If nothing changed for a given toot, its respective element won't be
touched.  If it got edited, it'll get new data.  If something new was
posted, a new element will be made.  And, assuming my CS 101 skills
hold up, the elements are sorted at creation time to fit in with
everything already on the page.

It's the sort of thing that might really just be overcomplicating things
just a skoshe, but I realized this might be a good idea when I noticed
that whenever the refresh timeout happens, anything unspoilered suddenly
became spoilered again, and then it kinda snowballed into other ideas,
and this was the result.